refactor: update imports to use consolidated types
